% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{all_bar_plot}
\alias{all_bar_plot}
\title{all_bar_plot function to plot all vars with bar}
\usage{
all_bar_plot(df, my_vars, nrow = NULL)
}
\arguments{
\item{df}{dataset}

\item{my_vars}{df with info about variables}

\item{nrow}{numer of rows}
}
\description{
all_bar_plot function to plot all vars with bar
}
