% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{assert_duplicates}
\alias{assert_duplicates}
\title{Assert if duplicates in data.frame}
\usage{
assert_duplicates(x, .var.name = checkmate::vname(x), add = NULL)
}
\description{
Assert if duplicates in data.frame
}
