% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_my_vars}
\alias{check_my_vars}
\title{Check my_vars}
\usage{
check_my_vars(my_vars)
}
\arguments{
\item{my_vars}{dataframe with mapping from variable names to variable families and labels. One row per variable.}
}
\description{
Check my_vars
}
\examples{
my_vars <- data.frame(
variable = c("var1", "var2",  "var1", NA),
family = c("fam1", "fam2", "fam1", NA),
short_label = c("label1", "label2", "label1", NA),
description = c("d1", "d2", "d1", NA),
stringsAsFactors = FALSE
)
\dontrun{check_my_vars(my_vars)}

}
