% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_country_graph.R
\name{ft_country_graphs}
\alias{ft_country_graphs}
\title{dashdash}
\usage{
ft_country_graphs(
  my_country_code,
  df,
  lab_q_x = 0.95,
  lab_q_y = 0.95,
  plot_title = ""
)
}
\arguments{
\item{my_country_code}{Three letter country code}

\item{df}{dataframe with corona cases containing outcome}

\item{lab_q_x}{quantile for labelling on x axis}

\item{lab_q_y}{quantile for labelling on y axis}

\item{plot_title}{optional plot title}
}
\description{
Make a .html dashboard. Key required inputs are datasets with 1. outcomes 2. variable interpretations 3. other text 4. GIS information.
}
