% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_means.R
\name{get_current_means}
\alias{get_current_means}
\title{current means}
\usage{
get_current_means(df, my_vars, recent = 1)
}
\arguments{
\item{df}{dataframe}

\item{my_vars}{df with info about variables}

\item{recent}{how many days previous to most recent date to include (in days)}
}
\description{
current means
}
