% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_plots.R
\name{het_bar_plot}
\alias{het_bar_plot}
\title{het_bar plot function to plot all vars with bar divided by id}
\usage{
het_bar_plot(df, my_vars)
}
\arguments{
\item{df}{dataset}

\item{my_vars}{df with info about variables}
}
\description{
het_bar plot function to plot all vars with bar divided by id
}
