% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_plots.R
\name{het_together}
\alias{het_together}
\title{het_together function to plot together plot by id}
\usage{
het_together(df, my_vars)
}
\arguments{
\item{df}{dataset}

\item{my_vars}{df with info about variables}
}
\description{
het_together function to plot together plot by id
}
