% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_disaggregates.R
\name{plot_disaggregates}
\alias{plot_disaggregates}
\title{plot disaggregates over time}
\usage{
plot_disaggregates(df, my_vars, pd = ggplot2::position_dodge(0.1))
}
\arguments{
\item{df}{dataframe}

\item{my_vars}{df with info about variables}

\item{pd}{ggplot dodge parameter}

\item{vars}{vector of variable names}
}
\description{
plot disaggregates over time
}
