% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregates.R
\name{plot_mov_avg}
\alias{plot_mov_avg}
\title{plot 3 day moving average}
\usage{
plot_mov_avg(df, my_vars, pd = ggplot2::position_dodge(0.1), switch = "y")
}
\arguments{
\item{df}{dataframe}

\item{my_vars}{df with info about variables}

\item{pd}{ggplot dodge parameter}
}
\description{
plot 3 day moving average
}
