% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_bar_plot.R
\name{row_function}
\alias{row_function}
\title{row fun}
\usage{
row_function(num)
}
\arguments{
\item{num}{number of rows to truncate}

\item{my_vars}{df with info about variables}
}
\description{
row fun
}
