% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{summary table}
\usage{
summary_table(df, my_vars)
}
\arguments{
\item{df}{dataframe}

\item{my_vars}{df with info about variables}

\item{date}{date variable}
}
\description{
summary table
}
