% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_bar_plot.R
\name{together_plot}
\alias{together_plot}
\title{together_plot one plot with relative freq}
\usage{
together_plot(df, my_vars)
}
\arguments{
\item{df}{dataset}

\item{my_vars}{df with info about variables}
}
\description{
together_plot one plot with relative freq
}
