% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashdash.R
\name{dashdash}
\alias{dashdash}
\title{dashdash}
\usage{
dashdash(
  output_file,
  my_vars,
  my_data,
  my_args,
  title = NULL,
  group = NULL,
  subtitle = NULL,
  author = NULL,
  add_maps = NULL,
  map_path = NULL,
  map_region = NULL,
  map_layer = NULL,
  scale_vars = NULL,
  pd_width = 0.1,
  ft_plot = NULL,
  country_code = NULL,
  ...
)
}
\arguments{
\item{output_file}{Path to where ouutput should be written, e.g. "docs/index.html"}

\item{my_vars}{dataframe with mapping from variable names to variable families and labels. One row per variable.}

\item{my_data}{dataframe with data on variables listed in `my_vars`. Should contain a `id` variable that connects with units in map files and a `date` variable.}

\item{my_args}{dataframe with arguments to customize a dashboard.}

\item{title}{Dashboard title}

\item{subtitle}{Dashboard subtitle}

\item{author}{String, names of authors.}

\item{map_path}{String, path to map shapefiles.}

\item{map_region}{String, map region.}

\item{map_layer}{String, map layer.}

\item{scale_vars}{Logical. Whether to scale_vars variabels before map plotting.}

\item{ft_plot}{Add a Financial Times type plot}

\item{country_code}{three letter country code used for Financial Times plot}

\item{switch}{ggplot2 argument: By default, the labels are displayed on the top and right of the plot. If "x", the top labels will be displayed to the bottom. If "y", the right-hand side labels will be displayed to the left. Can also be set to "both".}
}
\description{
Make a .html dashboard. Key required inputs are datasets with 1. outcomes 2. variable interpretations 3. other text 4. GIS information.
}
\details{
Note I had a problem with `Error: isTRUE(gpclibPermitStatus()) is not TRUE` which was solved
with `install.packages('rgeos', type='source')` and `install.packages('rgdal', type='source')`
}
\examples{
library(dplyr)

my_data <-
  expand_grid(id = c("Bo", "Bombali", "Bonthe"), date = as.Date(c("2020-04-18", "2020-04-19", "2020-04-20"))) \%>\%
  data.frame(stringsAsFactors = FALSE) \%>\%
    mutate(n = 3) \%>\%
    uncount(n) \%>\%
    dplyr::mutate(
    market_open	 = rbinom(n(), 1, .5),
    price_rice	 = rbinom(n(), 1, .5),
    aware	 = rbinom(n(), 1, .5),
    water	 = rbinom(n(), 1, .5))

my_vars <- data.frame(
  variable = c("market_open", "price_rice", "aware", "water"),
  family = c("Markets", "Markets", "Actions", "Actions"),
  short_label = c("Is market open?", "Price of a rice", "Aware of Covid19", "Access to water"),
  description = c("details on market open", "Price of a cup of rice", "Details on aware of Covid19", "Details on access to water"),
  stringsAsFactors = FALSE
)

my_args <- data.frame(
  para = "Intro text",
  datanote = "Study specific note about data source",
  map_path = "c:/temp/shapefiles",
  map_layer = "SLE_adm3",
  map_region =  "NAME_2",
  stringsAsFactors = FALSE
  )



dashdash(output_file = "dashtest.html",
         title = "title2",
         subtitle = "subtitle here",
         my_data = my_data,
         my_vars = my_vars,
         my_args = my_args)

}
