% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregates.R
\name{plot_aggregates}
\alias{plot_aggregates}
\title{plot aggregates over time}
\usage{
plot_aggregates(df, my_vars, pd = ggplot2::position_dodge(0.1))
}
\arguments{
\item{df}{dataframe}

\item{my_vars}{df with info about variables}

\item{pd}{ggplot dodge parameter}

\item{vars}{vector of variable names}
}
\description{
plot aggregates over time
}
